pipelines/bak_serialization_distribution/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "bak_serialization_distribution"
version = "0.0.2"
dependencies = [
    "azure-identity == 1.23.0",
    "azure-storage-blob == 12.27.0b1",
]

[project.optional-dependencies]
build = [
    "build == 0.10.0",
]
test = [
    "delta-spark == 2.4.0",
    "opencensus-ext-azure==1.1.9",
    "opencensus-ext-logging==0.1.1",
    "pyodbc == 4.0.39",
    "pyspark == 3.4.0",
    "pytest == 7.2.1",
    "pytest-mock == 3.11.1",
    "pymssql == 2.2.8",
    "pyspark-test == 0.2.0", 
    "testcontainers == 3.7.1",
    "SQLAlchemy == 2.0.19",
    "azure-identity == 1.23.0",
    "azure-storage-blob == 12.27.0b1",
]
lint = [
    "mypy == 1.0.0"
]

[tool.black]
line-length = 99

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opencensus.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyspark_test.*"
ignore_missing_imports = true
